Rename `sync_runtime` module to `runtime`
No need to have a distinguishing prefix anymore.
